requirements.txt
fastapi
uvicorn
ariadne
SPARQLWrapper
pyyaml